requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.32.3
